allow more proxies, and fail if this number is exceeded.
Former-commit-id: 15ef95aecade7eae75efe24e882f4bc67d8c0222

+++ src/cell.h
@@ -133,7 +133,7 @@ struct cell {
     int nodeID;
     
     /* Bit mask of the proxies this cell is registered with. */
-    unsigned int sendto;
+    unsigned long long int sendto;
     
     /* Pointer to this cell's packed representation. */
     struct pcell *pcell;
